requirements.txt
torch==1.7.0
flair==0.6.1.post1
seqeval==1.2.2
PyYAML==5.3.1